comm-mavlink-relay: Success without encryption

+++ apps/uart2/components/Echo2/src/echo2.c
@@ -185,7 +185,7 @@ int run(void) {
         // while (c != EOF) {
             // handle_char(c);
             // ringbuffer.buffer[ringbuffer.tail++] = c;
-            uint8_t tail;
+            uint32_t tail;
             ring_buffer_t *ringbuffer = (ring_buffer_t *) rb;
             tail = ringbuffer->tail;
             rb_acquire();
